requirements.txt
transformers==4.31.0
Pillow==10.0.0
Flask==2.3.2
torch==2.0.1